Move annotation LiteralParameter to SPI for use in other modules
