+++ presto-main-base/src/main/java/com/facebook/presto/operator/scalar/CharacterStringCasts.java
@@ -14,10 +14,10 @@
 package com.facebook.presto.operator.scalar;
 
 import com.facebook.presto.common.function.OperatorType;
+import com.facebook.presto.spi.function.LiteralParameter;
 import com.facebook.presto.spi.function.LiteralParameters;
 import com.facebook.presto.spi.function.ScalarOperator;
 import com.facebook.presto.spi.function.SqlType;
-import com.facebook.presto.type.LiteralParameter;
 import com.google.common.collect.ImmutableList;
 import io.airlift.slice.Slice;
 import io.airlift.slice.SliceUtf8;
